Parse and validate URLs in HTML
